Update news articles handlers and id shared type and fix enum bugs

+++ src/db/tables/orders.rs
@@ -9,7 +9,7 @@ pub async fn create_orders_table(client: Arc<Mutex<Client>>) -> Result<(), Box<d
     let create_users_table_sql = "
         CREATE TABLE IF NOT EXISTS orders (
         id BIGSERIAL PRIMARY KEY,
-        user_id INTEGER NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+        user_id BIGINT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
         order_date TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
         status VARCHAR(20) NOT NULL DEFAULT 'pending',
         total_amount NUMERIC(10, 2) NOT NULL
