+++ src/shared/types.rs
@@ -1,1 +1,1 @@
-pub type Id = i32;
+pub type Id = i64;
